requirements.txt
torch==1.13.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==0.13.1
torchmetrics==0.11.4
librosa==0.9.2
numpy==1.22.4
numba==0.55.2
SoundFile==0.10.3.post1
pyannote.audio==0.0.1
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
matplotlib==3.7.3
matplotlib-inline==0.1.6